+++ sprng-boot/lesson-32/lesson-32-task/src/main/java/az/atl/academy/lesson32task/controller/EmployeeController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("api/v1")
+@RequestMapping("api/v1/employees")
 public class EmployeeController {
 
     private final EmployeeService employeeService;
@@ -20,44 +20,34 @@ public class EmployeeController {
         this.employeeService = employeeService;
     }
 
-    @GetMapping("/departments")
-    public List<DepartmentDto> getDepartments() {
-        return employeeService.getDepartments();
-    }
-
-    @GetMapping("/departments/{id}")
-    public ResponseEntity<DepartmentDto> getDepartment(@PathVariable("id") Long departmentId) {
-        return employeeService.getDepartment(departmentId);
-    }
-
-    @PostMapping("/employees")
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public int insertEmployee(@RequestBody EmployeeDto employee) {
         return employeeService.insertEmployee(employee);
     }
 
-    @PostMapping("/departments")
-    public void insertNewDepartmentsBatch(@RequestBody DepartmentDto department1,
-                                          @RequestBody DepartmentDto department2) {
-        employeeService.insertNewDepartmentsBatch(department1, department2);
-    }
+//    @PostMapping("/departments")
+//    public void insertNewDepartmentsBatch(@RequestBody DepartmentDto department1,
+//                                          @RequestBody DepartmentDto department2) {
+//        employeeService.insertNewDepartmentsBatch(department1, department2);
+//    }
 
-    @GetMapping("/employees")
+    @GetMapping
     public List<EmployeeDto> getEmployees() {
         return employeeService.getEmployees();
     }
 
-    @GetMapping("/employees/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<EmployeeDto> getEmployee(@PathVariable("id") Long employeeId) {
         return employeeService.getEmployee(employeeId);
     }
 
-    @PutMapping("/employees/{id}")
+    @PutMapping("/{id}")
     public void updateEmployee(@PathVariable("id") Long employeeId, @RequestBody EmployeeDto employee) {
         employeeService.updateEmployee(employeeId, employee);
     }
 
-    @DeleteMapping("/employees/{id}")
+    @DeleteMapping("/{id}")
     public void deleteEmployee(@PathVariable("id") Long employeeId) {
         employeeService.deleteEmployee(employeeId);
     }
