+++ sprng-boot/lesson-32/lesson-32-task/src/main/java/az/atl/academy/lesson32task/repository/DepartmentRepository.java
@@ -1,0 +1,13 @@
+package az.atl.academy.lesson32task.repository;
+
+import az.atl.academy.lesson32task.dto.DepartmentDto;
+
+import org.springframework.http.ResponseEntity;
+
+import java.util.List;
+
+public interface DepartmentRepository {
+    List<DepartmentDto> getDepartments();
+
+    ResponseEntity<DepartmentDto> getDepartment(Long departmentId);
+}
