+++ sprng-boot/lesson-32/lesson-32-task/src/main/java/az/atl/academy/lesson32task/service/EmployeeService.java
@@ -3,6 +3,7 @@ package az.atl.academy.lesson32task.service;
 import az.atl.academy.lesson32task.dto.DepartmentDto;
 import az.atl.academy.lesson32task.dto.EmployeeDto;
 import az.atl.academy.lesson32task.repository.EmployeeRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -10,22 +11,10 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class EmployeeService {
 
     private final EmployeeRepository employeeRepository;
-
-    @Autowired
-    public EmployeeService(EmployeeRepository employeeRepository) {
-        this.employeeRepository = employeeRepository;
-    }
-
-    public List<DepartmentDto> getDepartments() {
-        return employeeRepository.getDepartments();
-    }
-
-    public ResponseEntity<DepartmentDto> getDepartment(Long departmentId) {
-        return employeeRepository.getDepartment(departmentId);
-    }
 
     public int insertEmployee(EmployeeDto employee) {
         return employeeRepository.insertEmployee(employee);
@@ -34,7 +23,6 @@ public class EmployeeService {
     public void insertNewDepartmentsBatch(DepartmentDto department1, DepartmentDto department2) {
         employeeRepository.insertNewDepartmentsBatch(department1, department2);
     }
-
 
     public void updateEmployee(Long employeeId, EmployeeDto employee) {
         employeeRepository.updateEmployee(employeeId, employee);
